% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsfy_performance.R
\name{clsfy_performance}
\alias{clsfy_performance}
\title{Classification performance measures}
\usage{
clsfy_performance(scores, true_cl, threshold = 0.5)
}
\arguments{
\item{scores}{probabilities or scores for class 1 (1 - prob of class 2);
scores are assumed to be in \eqn{[0, 1]} and high scores correspond to high
probability.}

\item{true_cl}{vector of true class labels.}

\item{threshold}{threshold for prediction, see \code{\link[jdify:predict.jdify]{jdify::predict.jdify()}}.}
}
\value{
A data.frame where each column corresponds to one value of \code{threshold}.
The corresponding values can be found with \code{attr(result, "threshold")}.
}
\description{
Calculates performance measures for a calssifier Assumes there are two
classes, and the first of \code{level(true_cl)} is to be predicted (the
"positive").
}
\examples{
# simulate training and test data
dat <- data.frame(x1 = rnorm(10), x2 = rbinom(10, 1, 0.3))
dat$cl <- c("A", "B")[round(pnorm(dat$x1 + dat$x2)) + 1]
dat_test <- data.frame(x1 = rnorm(10), x2 = rbinom(10, 1, 0.3))
dat_test$cl <- c("A", "B")[round(pnorm(dat_test$x1 + dat_test$x2)) + 1]

# fit density with cckde
fit_fun <- function(x) cctools::cckde(x)
eval_fun <- function(object, newdata) cctools::dcckde(newdata, object)
model <- jdify(cl ~ x1 + x2, dat, fit_fun, eval_fun)

probs <- predict(model, dat, what = "cprobs")  # conditional probabilities

# calculate performance measures
t <- seq(0, 1, by = 0.05)
perf <- clsfy_performance(scores = probs[, 1], dat[, 1], threshold = t)

# ROC plot
plot(t(perf[c("TPR", "FPR"), ]), type = "l", xlim = c(0, 1), ylim = c(0, 1))
abline(0, 1)

}
