% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsfy_performance.R
\name{clsfyr_rocplot}
\alias{clsfyr_rocplot}
\title{ROC plots}
\usage{
clsfyr_rocplot(x, cols = NULL, ltys = NULL)
}
\arguments{
\item{x}{object of class \code{"clsfyr_performance"} or a list of such objects.}

\item{cols}{vector of colors for the \code{"clsfyr_performance"} object.}

\item{ltys}{vector of line types for the `"clsfyr_performance"`` object.}
}
\description{
Produces a ROC plot for one or several \code{"clsfyr_performance"} objects.
}
