% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsfyr.R
\name{get_auc}
\alias{get_auc}
\title{Calculate the area under the curve (AUC)}
\usage{
get_auc(roc)
}
\arguments{
\item{roc}{a matrix or data frame with columns \code{"FPR"} and \code{"TPR"} containing
pairs of false and true positive rates.#'}
}
\description{
Calculate the area under the curve (AUC)
}
\examples{
# simulate training and test data
dat <- data.frame(
    cl = as.factor(rbinom(10, 1, 0.5)),
    x1 = rnorm(10),
    x2 = rbinom(10, 1, 0.3)
)

model <- jdify(cl ~ x1 + x2, data = dat)      # joint density fit
probs <- predict(model, dat, what = "probs")  # conditional probabilities

# calculate performance measures
assess_clsfyr(probs[, 1], dat[, 1] == 0, measure = c("ACC", "F1"))

# calculate area under the curve
FPR <- assess_clsfyr(probs[, 1], dat[, 1] == 0, measure = c("FPR"))$value
TPR <- assess_clsfyr(probs[, 1], dat[, 1] == 0, measure = c("TPR"))$value
get_auc(data.frame(FPR = FPR, TPR = TPR))

}
