% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.jdify.R
\name{predict.jdify}
\alias{predict.jdify}
\title{Predict method for jdify objects}
\usage{
\method{predict}{jdify}(object, newdata, what = "class", threshold = 0.5,
  ...)
}
\arguments{
\item{object}{an object of class \code{jdify}.}

\item{newdata}{a \code{data.frame} containing the predictors.}

\item{what}{either \code{"class"} for class predictions or \code{"probs"} for condition
probability estimates.}

\item{threshold}{threshold parameter for class prediction (\code{what = "class"});
predicts class \eqn{C} if \eqn{Pr(C) \ge threshold}.}

\item{...}{unused.}
}
\value{
A data frame containing the predictions.
}
\description{
Predict method for jdify objects
}
\examples{
# simulate training and test data
dat <- data.frame(
    cl = as.factor(rbinom(10, 1, 0.5)),
    x1 = rnorm(10),
    x2 = rbinom(10, 1, 0.3)
)

# fit density with cctools package
model <- jdify(cl ~ x1 + x2, data = dat, jd_method("cctools"))
pred <- predict(model, dat)                   # class predictions
probs <- predict(model, dat, what = "probs")  # conditional probabilities

}
