% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsfyr.R
\name{clsfyr_rocplot}
\alias{clsfyr_rocplot}
\title{ROC plots}
\usage{
clsfyr_rocplot(x, cols = NULL, ltys = NULL)
}
\arguments{
\item{x}{object of class \code{"clsfyr_performance"} or a list of such objects.}

\item{cols}{vector of colors for the \code{"clsfyr_performance"} object.}

\item{ltys}{vector of line types for the `"clsfyr_performance"`` object.}
}
\description{
Produces a ROC plot for one or several \code{"clsfyr_performance"} objects.
}
\examples{
# simulate training and test data
dat <- data.frame(
    cl = as.factor(rbinom(10, 1, 0.5)),
    x1 = rnorm(10),
    x2 = rbinom(10, 1, 0.3)
)

model <- jdify(cl ~ x1 + x2, data = dat)       # joint density fit
probs <- predict(model, dat, what = "cprobs")  # conditional probabilities

# calculate performance measures
t <- seq(0, 1, by = 0.05)
perf <- clsfyr_performance(scores = probs[, 1], dat[, 1], threshold = t)

# ROC plot
plot(t(perf[c("TPR", "FPR"), ]), type = "l", xlim = c(0, 1), ylim = c(0, 1))
abline(0, 1)

}
