% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_jdify.R
\name{cv_jdify}
\alias{cv_jdify}
\title{Cross validation for joint density classification}
\usage{
cv_jdify(formula, data, fit_fun = function(x, ...) NULL, eval_fun = predict,
  cc = FALSE, folds = 10, cores = 1, ...)
}
\arguments{
\item{formula}{an object of class "formula"; same as \code{\link[stats:lm]{stats::lm()}}.}

\item{data}{matrix, data frame, list or environment (or object coercible by
\code{\link[base:as.data.frame]{base::as.data.frame()}}) containing the variables in the model.}

\item{fit_fun}{a function of type \code{function(x, ...)} that fits a joint
density model on a data matrix \code{x}. The \code{...} can be used for passing
additional parameters.}

\item{eval_fun}{a function of type \code{function(object, newdata, ...)} that
takes an object fitted by \code{fit_fun} and evaluates the density estimate on
\code{newdata}.}

\item{cc}{if \code{TRUE}, discrete variables (and the class indicator) are made
continuous with \code{\link[cctools:cont_conv]{cctools::cont_conv()}}; only use \code{FALSE} when your \code{fit_fun}
can handle discrete variables.}

\item{folds}{number of folds.}

\item{cores}{number of cores for parallelized cross validation (based on
\code{\link[foreach:foreach]{foreach::foreach()}}).}

\item{...}{unused.}
}
\description{
Performs k-fold cross validation for a \code{jdify} object.
}
