% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdify-package.R
\docType{package}
\name{jdify-package}
\alias{jdify-package}
\title{Joint density classification}
\description{
Implements classifiers based on the joint density of the predictors
and the class variable. Several methods for joint density estimation
can be used.
}
\references{
Nagler, T. (2017). \emph{A generic approach to nonparametric function
estimation with mixed data.} \href{https://arxiv.org/abs/1704.07457}{arXiv:1704.07457}
}
\author{
Thomas Nagler
}
\keyword{package}
