% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdify.R
\name{jdify}
\alias{jdify}
\title{Joint density classification}
\usage{
jdify(formula, data, jd_method = "cctools", ...)
}
\arguments{
\item{formula}{an object of class "formula"; same as \code{\link[stats:lm]{stats::lm()}}.}

\item{data}{matrix, data frame, list or environment (or object coercible by
\code{\link[base:as.data.frame]{base::as.data.frame()}}) containing the variables in the model.}

\item{jd_method}{an object of class \code{"jd_method"} or a character string
for built-in methods, see \code{\link[=jd_method]{jd_method()}}.}

\item{...}{additional parameters passed to \code{fit_fun}.}
}
\value{
An object of class \code{jdify}.
}
\description{
Fits a generic joint density model for the predictors and class variables.
}
\details{
For predicting the class or conditional probabilities, see,
\code{\link[=predict.jdify]{predict.jdify()}}.
}
\examples{
# simulate training and test data
dat <- data.frame(
    cl = as.factor(rbinom(10, 1, 0.5)),
    x1 = rnorm(10),
    x2 = rbinom(10, 1, 0.3)
)

model <- jdify(cl ~ x1 + x2, dat)              # fitted joint density
pred <- predict(model, dat)                    # class predictions
probs <- predict(model, dat, what = "probs")   # conditional probabilities

# in-sample performance
assess_clsfyr(probs[, 1], dat[, 1] == 0, measure = c("ACC", "F1"))

}
\seealso{
\code{\link[=predict.jdify]{predict.jdify()}}, \code{\link[=assess_clsfyr]{assess_clsfyr()}}
}
