% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdify.R
\name{predict.jdify}
\alias{predict.jdify}
\title{Predict method for jdify objects}
\usage{
\method{predict}{jdify}(object, newdata, what = "class", threshold = 0.5,
  ...)
}
\arguments{
\item{object}{an object of class \code{jdify}.}

\item{newdata}{a \code{data.frame} containing the predictors.}

\item{what}{either \code{"class"} for class predictions or \code{"probs"} for condition
probability estimates.}

\item{threshold}{threshold parameter for class prediction (\code{what = "class"});
predicts the first element of \code{levels(class)} if \code{prob1 > threshold}, the
second otherwise.}

\item{...}{unused.}
}
\value{
A data frame containing the predictions.
}
\description{
Predict method for jdify objects
}
\examples{
# simulate training and test data
dat <- data.frame(x1 = rnorm(10), x2 = rbinom(10, 1, 0.3))
dat$cl <- c("A", "B")[round(pnorm(dat$x1 + dat$x2)) + 1]
dat_test <- data.frame(x1 = rnorm(10), x2 = rbinom(10, 1, 0.3))
dat_test$cl <- c("A", "B")[round(pnorm(dat_test$x1 + dat_test$x2)) + 1]

# fit density with cckde
fit_fun <- function(x) cctools::cckde(x)
eval_fun <- function(object, newdata) cctools::dcckde(newdata, object)
model <- jdify(cl ~ x1 + x2, dat, fit_fun, eval_fun)

pred <- predict(model, dat)                    # class predictions
probs <- predict(model, dat, what = "cprobs")  # conditional probabilities

}
